added a more stylish, better looking logo with the name of the project to the layout.
placed the signup view in a different view from the login view.

added shortcuts for teachers, students, signup views in the layout.

added an exit button that takes back to the login view on the layout.

Front-end validation with javascript now also validates whether user and pass are empty or if only one of them  is.